merge desktopbrightness into brightness
Detect if we're running on a laptop. If so, use classic brightnessctl,
but omit sudo because we prefer to setuid it for consistency and ease of
use. If we're not running on a laptop, use ddcutil to send brightness
commands to the monitors, with the monitor model as env var
(`desktop_monitor`), a fallback to the first monitor found, and a cache
to speed up monitor discovery. The cache allows us to have the ~same
speed as before, but now we can actually match a model name with a bus
number without resorting to ddcutil making 2 calls (a call being
~500ms).